Add passport req.logIn
In the custom callback section of the Passport docs: "The callback can
use the arguments supplied to handle the authentication result as
desired. Note that when using a custom callback, it becomes the
application's responsibility to establish a session (by calling
req.login()) and send a response."

Sessions don't appear to work as currently implemented, this adds the
req.logIn function.